style(emotion): upgrade emotion and use cleaner styling methodology (#352)
## Main changes

### Allow for cascading styles

Previously, drilling styling through a series of `configurableCss` functions was required to provide overrides. This can be replaced now with an array of styles. The styles are merged together w/ latter styling being preferred to the former.


```tsx
const baseStyle = css`
  font-size: 3rem;
  color: blue;
`

const override = css`
  font-size: 2rem;
`

// in some component...

  <h1 css={[baseStyle, override]} aria-label={content}>
    {content}
  </h1>
```

Using the above example, the `h1` component would be styled with the merged value of:

```css
  font-size: 2rem;
  color: blue;
```

### Update emotion

Use emotion 11 and properly configure it throughout the template.

Co-authored-by: Julian Coy <37161095+eXamadeus@users.noreply.github.com>